docs: change doc structure
