+++ stories/utils/initialRois.ts
